requirements.txt
fastapi
uvicorn
python-dotenv
requests
pydantic
pandas
sqlalchemy
psycopg2-binary  
tabulate         
langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-google-genai
google-generativeai
langchain-chroma
chromadb
langchain-huggingface
sentence-transformers
langgraph
streamlit